+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * CytoExploreR can now read in spillover matrices from `.mtx` files using the new internal `read_from_mtx()` function.
 * CytoExplreR can now be run without using a gatingTemplate by setting `cyto_gatingTemplate_active(FALSE)` at the beginning of the session.
+* `cyto_plot_map()` now supports plotting gates through the `alias` argument to maintain consistency with `cyto_plot()`.
 
 # CytoExploreR 2.0.2